big directory refactor
