+++ .gitignore
@@ -3,6 +3,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
+*__pycache__*
 
 # C extensions
 *.so
@@ -129,6 +130,9 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
+# Misc.
+*.vscode*
+*.vscode/*
 TransPath_data
 wandb
 map_data
